article에 memberId 정보 추가

+++ src/main/java/com/koreaIT/demo/controller/UsrArticleController.java
@@ -28,6 +28,10 @@ public class UsrArticleController {
 	@ResponseBody
 	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
 		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-0", "로그인 후 이용해주세요.");
+		}
+		
 		if(Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요.");
 		}
@@ -36,7 +40,9 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요.");
 		}
 		
-		articleService.writeArticle(title, body);
+		int memberId = (int) httpSession.getAttribute("loginedMemberId");
+		
+		articleService.writeArticle(title, body, memberId);
 		
 		int id = articleService.getLastInsertId();
 		
@@ -65,11 +71,20 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession httpSession, int id) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-0", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		if (!httpSession.getAttribute("loginedMemberId").equals(article.getMemberId())) {
+			return ResultData.from("F-2", "게시물 삭제 권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
@@ -79,11 +94,20 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-0", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		if (!httpSession.getAttribute("loginedMemberId").equals(article.getMemberId())) {
+			return ResultData.from("F-2", "게시물 수정 권한이 없습니다.");
 		}
 		
 		articleService.modifyArticle(id, title, body);
